feature(Authentication): refactor all dtos
