+++ src/main/kotlin/com/jlima/bookstoremanager/controller/author/AuthorController.kt
@@ -1,6 +1,6 @@
 package com.jlima.bookstoremanager.controller.author
 
-import com.jlima.bookstoremanager.dto.AuthorDTO
+import com.jlima.bookstoremanager.dto.author.AuthorDTO
 import com.jlima.bookstoremanager.dto.response.CustomMessageResponse
 import com.jlima.bookstoremanager.dto.response.PaginationResponse
 import com.jlima.bookstoremanager.service.AuthorService
