+++ src/main/kotlin/com/jlima/bookstoremanager/controller/publisher/PublisherController.kt
@@ -1,7 +1,7 @@
 package com.jlima.bookstoremanager.controller.publisher
 
 import com.jlima.bookstoremanager.controller.BaseController
-import com.jlima.bookstoremanager.dto.PublisherDTO
+import com.jlima.bookstoremanager.dto.publisher.PublisherDTO
 import com.jlima.bookstoremanager.dto.response.CustomMessageResponse
 import com.jlima.bookstoremanager.dto.response.PaginationResponse
 import com.jlima.bookstoremanager.service.PublisherService
