+++ src/main/kotlin/com/jlima/bookstoremanager/dto/auth/AuthResponse.kt
@@ -1,4 +1,4 @@
-package com.jlima.bookstoremanager.dto.response
+package com.jlima.bookstoremanager.dto.auth
 
 data class AuthResponse(
     val accessToken: String
