+++ src/main/kotlin/com/jlima/bookstoremanager/service/AuthorService.kt
@@ -1,6 +1,6 @@
 package com.jlima.bookstoremanager.service
 
-import com.jlima.bookstoremanager.dto.AuthorDTO
+import com.jlima.bookstoremanager.dto.author.AuthorDTO
 import com.jlima.bookstoremanager.dto.response.PaginationResponse
 import com.jlima.bookstoremanager.dto.response.toPaginationResponse
 import com.jlima.bookstoremanager.exception.model.AvailableEntities
