+++ src/test/kotlin/com/jlima/bookstoremanager/controller/author/AuthorControllerTest.kt
@@ -1,7 +1,7 @@
 package com.jlima.bookstoremanager.controller.author
 
 import com.jlima.bookstoremanager.config.security.JwtAuthenticationEntrypoint
-import com.jlima.bookstoremanager.dto.AuthorDTO
+import com.jlima.bookstoremanager.dto.author.AuthorDTO
 import com.jlima.bookstoremanager.dto.response.PaginationResponse
 import com.jlima.bookstoremanager.exception.model.AvailableEntities
 import com.jlima.bookstoremanager.exception.model.BusinessEmptyResponseException
